feat(portal): grouped sidebars, profile tabs, sidebar avatar photo
Both sidebars are regrouped with dividers. Member: Panel · Anuncios |
Actividades · Reservar | Mis actividades · Mis reservas · Mis recibos ·
Mi tarjeta. Admin: Panel | Socios · Grupos · Actividades · Espacios ·
Escanear | Recibos · Mandatos · Remesas · Facturación · Resumen anual |
Comunicaciones | Configuración. Empty groups collapse (feature flags).

Payment method leaves the nav: its editor moves to
features/members/my-payment-method and becomes the third tab on the
profile page (Editar perfil · Campos personalizados · Forma de pago,
lazy); /payment-method redirects to /profile?tab=payment for bookmarks.
The custom-fields card sheds its Card chrome for tab use and says so when
no fields are visible instead of rendering blank. Member-info header
tightened.

Sidebar avatar bug: the footer Avatar rendered initials only — /me now
exposes person.photo_url, and the sidebar renders AvatarImage with the
member-photo-upload cache-buster idiom, falling back to initials.

Also ignores the runtime /storage/ uploads dir at the repo root.
