+++ frontend/features/custom-fields/components/my-custom-fields-card.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { useTranslations } from "next-intl";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { FormSkeleton } from "@/components/ui/skeletons";
 import { useSettings } from "@/features/settings/hooks/use-settings";
 import {
@@ -11,10 +10,11 @@ import {
 } from "../hooks/use-custom-fields";
 import { CustomFieldsForm } from "./custom-fields-form";
 
-/** The member's own custom fields, on their profile page.
+/** The member's own custom fields — a tab on their profile page.
  *  Without this, a field set to member-writable would have nowhere to be
- *  written — members can't reach the member detail page. */
-export function MyCustomFieldsCard() {
+ *  written — members can't reach the member detail page. The tab trigger is
+ *  gated on the feature flag by the profile page; this renders content only. */
+export function MyCustomFields() {
   const t = useTranslations();
   const { data: settings } = useSettings();
   const enabled = Boolean(settings?.features?.custom_profile_fields);
@@ -25,28 +25,24 @@ export function MyCustomFieldsCard() {
     useMyCustomFields(enabled);
   const updateMutation = useUpdateMyCustomFields();
 
-  // Nothing to show when the feature is off, or when every field is hidden
-  // from this member.
   if (!enabled) return null;
-  if (!loadingDefinitions && !definitions.length) return null;
+  if (loadingDefinitions || loadingValues) return <FormSkeleton fields={2} />;
+  // Every field hidden from this member — say so rather than a blank panel.
+  if (!definitions.length)
+    return (
+      <p className="text-sm text-muted-foreground">
+        {t("profileFields.noFieldsForMember")}
+      </p>
+    );
 
   return (
-    <Card>
-      <CardHeader className="py-3 px-4">
-        <CardTitle className="text-base">{t("profileFields.tabLabel")}</CardTitle>
-      </CardHeader>
-      <CardContent className="px-4 pb-4 pt-0">
-        {loadingDefinitions || loadingValues ? (
-          <FormSkeleton fields={2} />
-        ) : (
-          <CustomFieldsForm
-            definitions={definitions}
-            values={values}
-            onSave={(next) => updateMutation.mutateAsync(next)}
-            isSaving={updateMutation.isPending}
-          />
-        )}
-      </CardContent>
-    </Card>
+    <div className="max-w-2xl">
+      <CustomFieldsForm
+        definitions={definitions}
+        values={values}
+        onSave={(next) => updateMutation.mutateAsync(next)}
+        isSaving={updateMutation.isPending}
+      />
+    </div>
   );
 }
